add: colored prompt with folder name
